requirements.txt
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.0
Flask==1.0.3
SQLAlchemy==1.3.5
google-cloud-storage==1.16.1
gunicorn==19.9.0
Wand==0.5.7
numpy==1.17.2
opencv-python==4.1.1.26
psycopg2-binary==2.8.3
